requirements.txt
# MCP Procurement Airlock Server
# Python 3.10+
mcp>=1.26.0
PyGithub>=2.5.0
requests>=2.28.0
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
defusedxml>=0.7.0
tabulate>=0.9.0
python-dotenv>=1.0.0
